Categorize and Modularize 1. Separate concerns into various sub-modules and integrate them into web-agent-module. 2. Abstract method calling, and only allow webAgentService to be publicily used.
