+++ src/app.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { PuppeteerModule } from './puppeteer/puppeteer.module';
 import { AirtableModule } from './airtable/airtable.module';
 import { DataLayerCheckerModule } from './data-layer-checker/data-layer-checker.module';
 import { GtmOperatorModule } from './gtm-operator/gtm-operator.module';
+import { WebAgentModule } from './web-agent/web-agent.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot(),
-    PuppeteerModule,
+    WebAgentModule,
     AirtableModule,
     DataLayerCheckerModule,
     GtmOperatorModule,
