+++ src/data-layer-checker/data-layer-checker.module.ts
@@ -3,11 +3,11 @@ import { Module } from '@nestjs/common';
 import { DataLayerCheckerController } from './data-layer-checker.controller';
 import { DataLayerCheckerService } from './data-layer-checker.service';
 import { HttpModule } from '@nestjs/axios';
-import { PuppeteerModule } from '../puppeteer/puppeteer.module';
 import { GtmOperatorModule } from '../gtm-operator/gtm-operator.module';
+import { WebAgentModule } from '../web-agent/web-agent.module';
 
 @Module({
-  imports: [HttpModule, AirtableModule, PuppeteerModule, GtmOperatorModule],
+  imports: [HttpModule, AirtableModule, GtmOperatorModule, WebAgentModule],
   controllers: [DataLayerCheckerController],
   providers: [DataLayerCheckerService],
 })
