+++ src/utilities/utilities.ts
@@ -1,0 +1,5 @@
+export function chunk(array: any[], chunkSize: number) {
+  return Array.from({ length: Math.ceil(array.length / chunkSize) }, (_, i) =>
+    array.slice(i * chunkSize, i * chunkSize + chunkSize),
+  );
+}
