+++ src/web-agent/action/action.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from '@nestjs/common';
+
+@Controller('action')
+export class ActionController {}
