+++ src/web-agent/action/action.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { ActionService } from './action.service';
+
+describe('ActionService', () => {
+  let service: ActionService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [ActionService],
+    }).compile();
+
+    service = module.get<ActionService>(ActionService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
