+++ src/web-agent/analysis/analysis.controller.ts
@@ -1,0 +1,6 @@
+import { Controller } from '@nestjs/common';
+
+@Controller('analysis')
+export class AnalysisController {
+  constructor() {}
+}
