+++ src/web-agent/analysis/analysis.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { AnalysisController } from './analysis.controller';
+import { AnalysisService } from './analysis.service';
+import { PuppeteerModule } from '../puppeteer/puppeteer.module';
+
+@Module({
+  imports: [PuppeteerModule],
+  controllers: [AnalysisController],
+  providers: [AnalysisService],
+  exports: [AnalysisService],
+})
+export class AnalysisModule {}
