+++ src/web-agent/analysis/analysis.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { AnalysisService } from './analysis.service';
+
+describe('AnalysisService', () => {
+  let service: AnalysisService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [AnalysisService],
+    }).compile();
+
+    service = module.get<AnalysisService>(AnalysisService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
