+++ src/web-agent/puppeteer/puppeteer.module.ts
@@ -1,10 +1,7 @@
 import { Module } from '@nestjs/common';
-import { PuppeteerController } from './puppeteer.controller';
 import { PuppeteerService } from './puppeteer.service';
 
 @Module({
-  imports: [],
-  controllers: [PuppeteerController],
   providers: [PuppeteerService],
   exports: [PuppeteerService],
 })
