+++ src/web-agent/utilities/utilities.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { UtilitiesService } from './utilities.service';
+
+@Module({
+  providers: [UtilitiesService],
+  exports: [UtilitiesService],
+})
+export class UtilitiesModule {}
