+++ src/web-agent/utilities/utilities.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { UtilitiesService } from './utilities.service';
+
+describe('UtilitiesService', () => {
+  let service: UtilitiesService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [UtilitiesService],
+    }).compile();
+
+    service = module.get<UtilitiesService>(UtilitiesService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
